+++ src/main/java/aed/hibernate/Stock.java
@@ -14,23 +14,23 @@ public class Stock {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "codproducto") 
-    private long codProducto;
+    private int codProducto;
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "codTienda")
-	private long codTienda;
+	private int codTienda;
 	@Column(name = "unidades")
 	private int unidades;
-	public long getCodProducto() {
+	public int getCodProducto() {
 		return codProducto;
 	}
-	public void setCodProducto(long codProducto) {
+	public void setCodProducto(int codProducto) {
 		this.codProducto = codProducto;
 	}
-	public long getCodTienda() {
+	public int getCodTienda() {
 		return codTienda;
 	}
-	public void setCodTienda(long codTienda) {
+	public void setCodTienda(int codTienda) {
 		this.codTienda = codTienda;
 	}
 	public int getUnidades() {
